Revert "Merge pull request #54 from ramey-plivo/powerpack"
This reverts commit b8809a9c737191662e86288caf2e88d08b2bae95, reversing
changes made to e531c04f4ff81f176ab3a9e09711c209e0dc9079.
