+++ CHANGELOG.md
@@ -1,15 +1,5 @@
 # Change Log
 
-## [4.2.0a3](https://github.com/plivo/plivo-python/tree/v4.2-alpha3) (2018-08-08)
-- Support python3.7
-- Add PHLO run function
-
-## [4.2.0a2](https://github.com/plivo/plivo-python/tree/v4.2-alpha2) (2018-08-03)
-- Fix PHLO Url and component name
-
-## [4.2.0a1](https://github.com/plivo/plivo-python/tree/v4.2-alpha1) (2018-08-03)
-- Add PHLO support in alpha release
-
 ## [4.1.1](https://github.com/plivo/plivo-python/tree/v4.1.1) (2018-06-29)
 - Added limit & offset parameters to endpoint list api
