+++ plivo/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import exceptions
 from .rest import Client as RestClient
-from .rest import phlo
 from . import xml as plivoxml
