--- plivo/rest/phlo.py
@@ -1,1 +1,0 @@
-from .phlo_client import PhloClient as RestClient
