+++ plivo/version.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-__version__ = '4.2.0a3'
+__version__ = '4.1.1'
